+++ src/LinkDotNet.LinqSIMDExtensions/Min.cs
@@ -29,6 +29,14 @@ public static partial class LinqSIMDExtensions
     }
 
     /// <summary>
+    /// Retrieves the minimum value of the memory.
+    /// </summary>
+    /// <returns>The max as type T.</returns>
+    public static T Min<T>(this Memory<T> memory)
+        where T : unmanaged, IMinMaxValue<T>, INumber<T>
+        => Min(memory.Span);
+
+    /// <summary>
     /// Retrieves the minimum value of the span.
     /// </summary>
     public static T Min<T>(this Span<T> span)