+++ tests/LinkDotNet.LinqSIMDExtensions.Tests/SumTests.cs
@@ -52,4 +52,14 @@ public class SumTests
 
         sum.ShouldBe(55);
     }
+
+    [Fact]
+    public void GivenMemory_WhenRetrievingSum_ThenTheCorrectSumIsReturned()
+    {
+        var numbers = new[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+
+        var sum = numbers.AsMemory().Sum();
+
+        sum.ShouldBe(55);
+    }
 }